requirements.txt
PyGithub>=2.1.0
anthropic>=0.34.0
openai>=1.40.0
PyYAML>=6.0
